Made the url for the current day reload to an url with the date embedded into it.
That way when we look at the page the next day
(meaning past midnight) we don't have to remember
to reload the page. Any effect on it, will be on
the correct date.
The reloading does involve a flicker/waste of time
though, but it seems worth it to me.
It would be good to make this optional though.

Signed-off-by: John-Gee <john.ettedgui@gmail.com>